# docs/openapi.json
{
 "openapi": "3.0.1",
 "info": {
  "title": "OpenAPI definition",
  "version": "v0"
 },
 "servers": [
  {
   "url": "http://localhost:18090",
   "description": "Generated server url"
  }
 ],
 "paths": {
...
  "/api/v1/subscriptions": {
   "post": {
    "tags": [
     "api-controller"
    ],
    "operationId": "subscribe",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SubRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "*/*": {
        "schema": {
         "type": "object",
         "additionalProperties": {
          "type": "object"
         }
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/messages": {
   "post": {
    "tags": [
     "api-controller"
    ],
    "summary": "Demander l'envoi d'un MT (idempotent via clientRef)",
    "operationId": "send",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SendRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "*/*": {
        "schema": {
         "$ref": "#/components/schemas/MessageView"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/content/{serviceId}/links": {
   "post": {
    "tags": [
     "api-controller"
    ],
    "operationId": "contentLink",
    "parameters": [
     {
      "name": "serviceId",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "format": "int64"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/LinkRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "*/*": {
        "schema": {
         "$ref": "#/components/schemas/MessageView"
        }
       }
      }
     }
    }
   }
  },
...
  "/admin/services": {
   "get": {
    "tags": [
     "admin-controller"
    ],
    "operationId": "services_2",
    "parameters": [
     {
      "name": "page",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "format": "int32"
      }
     },
     {
      "name": "size",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "format": "int32"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "*/*": {
        "schema": {
         "type": "array",
         "items": {
          "type": "object",
          "additionalProperties": {
           "type": "object"
          }
         }
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "admin-controller"
    ],
    "operationId": "createService",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ServiceReq"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "*/*": {
        "schema": {
         "type": "object",
         "additionalProperties": {
          "type": "object"
         }
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/messages/{id}": {
   "get": {
    "tags": [
     "api-controller"
    ],
    "operationId": "get",
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "*/*": {
        "schema": {
         "$ref": "#/components/schemas/MessageView"
        }
       }
      }
     }
    }
   }
  },
...
  "/admin/routing/resolve": {
   "get": {
    "tags": [
     "routing-admin-controller"
    ],
    "operationId": "resolve",
    "parameters": [
     {
      "name": "msisdn",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "*/*": {
        "schema": {
         "type": "object",
         "additionalProperties": {
          "type": "object"
         }
        }
       }
      }
     }
    }
   }
  },
...
  "/admin/operators": {
   "get": {
    "tags": [
     "admin-controller"
    ],
    "operationId": "operators",
    "parameters": [
     {
      "name": "page",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "format": "int32"
      }
     },
     {
      "name": "size",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "format": "int32"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "*/*": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/Operator"
         }
        }
       }
      }
     }
    }
   }
  },
  "/admin/messages": {
   "get": {
    "tags": [
     "reporting-query-controller"
    ],
    "operationId": "messages",
    "parameters": [
     {
      "name": "id",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string"
      }
     },
     {
      "name": "msisdn",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string"
      }
     },
     {
      "name": "operator",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string"
      }
     },
     {
      "name": "shortCode",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string"
      }
     },
     {
      "name": "status",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "enum": [
        "PENDING",
        "SUBMITTED",
        "DELIVERED",
        "EXPIRED",
        "UNDELIVERABLE",
        "REJECTED",
        "UNKNOWN",
        "FAILED"
       ]
      }
     },
     {
      "name": "from",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "format": "date-time"
      }
     },
     {
      "name": "to",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "format": "date-time"
      }
     },
     {
      "name": "page",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "format": "int32"
      }
     },
     {
      "name": "size",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "format": "int32"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "*/*": {
        "schema": {
         "type": "array",
         "items": {
          "type": "object",
          "additionalProperties": {
           "type": "object"
          }
         }
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/subscriptions/{id}": {
   "delete": {
    "tags": [
     "api-controller"
    ],
    "operationId": "unsubscribe",
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "format": "int64"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "OK"
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "SubRequest": {
    "required": [
     "consentProof",
     "msisdn",
     "serviceId"
    ],
    "type": "object",
    "properties": {
     "msisdn": {
      "type": "string"
     },
     "serviceId": {
      "type": "integer",
      "format": "int64"
     },
     "consentProof": {
      "type": "string"
     }
    }
   },
   "SendRequest": {
    "required": [
     "text",
     "to"
    ],
    "type": "object",
    "properties": {
     "to": {
      "type": "string"
     },
     "text": {
      "type": "string"
     },
     "sender": {
      "type": "string"
     },
     "serviceId": {
      "type": "integer",
      "format": "int64"
     },
     "operator": {
      "type": "string"
     },
     "priority": {
      "type": "string",
      "enum": [
       "TRANSACTIONAL",
       "CONFIRMATION",
       "BULK"
      ]
     },
     "clientRef": {
      "type": "string"
     },
     "scheduleAt": {
      "type": "string",
      "format": "date-time"
     }
    }
   },
   "MessageView": {
    "type": "object",
    "properties": {
     "id": {
      "type": "string"
     },
     "clientRef": {
      "type": "string"
     },
     "to": {
      "type": "string"
     },
     "status": {
      "type": "string"
     },
     "rawStatus": {
      "type": "string"
     },
     "segments": {
      "type": "integer",
      "format": "int32"
     },
     "encoding": {
      "type": "string"
     },
     "history": {
      "type": "array",
      "items": {
       "type": "object",
       "additionalProperties": {
        "type": "object"
       }
      }
     }
    }
   },
   "LinkRequest": {
    "required": [
     "msisdn"
    ],
    "type": "object",
    "properties": {
     "msisdn": {
      "type": "string"
     },
     "code": {
      "type": "string"
     }
    }
   },
...
   "Operator": {
    "type": "object",
    "properties": {
     "id": {
      "type": "integer",
      "format": "int64"
     },
     "code": {
      "type": "string"
     },
     "name": {
      "type": "string"
     },
     "msisdnPrefixes": {
      "type": "string"
     },
     "jasminConnector": {
      "type": "string"
     },
     "dlrBillingRule": {
      "type": "string",
      "enum": [
       "ON_SUBMITTED",
       "ON_DELIVERED"
      ]
     },
     "maxTps": {
      "type": "integer",
      "format": "int32"
     },
     "status": {
      "type": "string"
     },
     "configApplied": {
      "type": "boolean"
     }
    }
   },
...
   "ServiceReq": {
    "type": "object",
    "properties": {
     "name": {
      "type": "string"
     },
     "type": {
      "type": "string",
      "enum": [
       "VOTE",
       "QUIZ",
       "PREMIUM_CONTENT",
       "SUBSCRIPTION",
       "ALERT"
      ]
     },
     "partnerId": {
      "type": "integer",
      "format": "int64"
     },
     "shortCodeId": {
      "type": "integer",
      "format": "int64"
     },
     "regulated": {
      "type": "boolean"
     },
     "consentMode": {
      "type": "string",
      "enum": [
       "SIMPLE_OPT_IN",
       "DOUBLE_OPT_IN",
       "API_ACTIVATION"
      ]
     },
     "opensAt": {
      "type": "string",
      "format": "date-time"
     },
     "closesAt": {
      "type": "string",
      "format": "date-time"
     },
     "maxActionsPerMsisdn": {
      "type": "integer",
      "format": "int32"
     },
     "defaultLang": {
      "type": "string"
     },
     "replyOk": {
      "type": "string"
     },
     "replyStop": {
      "type": "string"
     },
     "replyHelp": {
      "type": "string"
     },
     "replyLimit": {
      "type": "string"
     },
     "replyClosed": {
      "type": "string"
     },
     "maxTps": {
      "type": "integer",
      "format": "int32"
     }
    }
   },
...
  }
 }
}